fix: merged terminal and terminalInstance into one type
