+++ src/types/index.ts
@@ -1,4 +1,3 @@
 export * from './terminal'
 export * from './terminalCommand'
-export * from './terminalInstance'
 export * from './terminalSettings'
