+++ src/types/terminalCommand.ts
@@ -1,8 +1,8 @@
-import { Terminal } from './terminal'
+import { TerminalInstance } from './terminal'
 
 export type TerminalCommand = {
   name: string
   description: string
   argDescriptions?: string[]
-  func: (terminal: Terminal, ...args: string[]) => void
+  func: (terminal: TerminalInstance, ...args: string[]) => void
 }
